Project Setup and Basic Mechanics
I have set up the project and added a basic movement, cinemachine freelook camera code and a simple build mode.

The build mode can be toggled with "B", The build block can be moved with WASDQ and E or arrows keys, right shift and control, blocks can also be placed and overlapped with space.